+++ gems/aws-sdk-costexplorer/lib/aws-sdk-costexplorer/endpoints.rb
@@ -389,6 +389,20 @@ module Aws::CostExplorer
       end
     end
 
+    class ListSavingsPlansPurchaseRecommendationGeneration
+      def self.build(context)
+        unless context.config.regional_endpoint
+          endpoint = context.config.endpoint.to_s
+        end
+        Aws::CostExplorer::EndpointParameters.new(
+          region: context.config.region,
+          use_dual_stack: context.config.use_dualstack_endpoint,
+          use_fips: context.config.use_fips_endpoint,
+          endpoint: endpoint,
+        )
+      end
+    end
+
     class ListTagsForResource
       def self.build(context)
         unless context.config.regional_endpoint
@@ -404,6 +418,20 @@ module Aws::CostExplorer
     end
 
     class ProvideAnomalyFeedback
+      def self.build(context)
+        unless context.config.regional_endpoint
+          endpoint = context.config.endpoint.to_s
+        end
+        Aws::CostExplorer::EndpointParameters.new(
+          region: context.config.region,
+          use_dual_stack: context.config.use_dualstack_endpoint,
+          use_fips: context.config.use_fips_endpoint,
+          endpoint: endpoint,
+        )
+      end
+    end
+
+    class StartSavingsPlansPurchaseRecommendationGeneration
       def self.build(context)
         unless context.config.regional_endpoint
           endpoint = context.config.endpoint.to_s