+++ gems/aws-sdk-quicksight/lib/aws-sdk-quicksight/client.rb
@@ -9279,7 +9279,7 @@ module Aws::QuickSight
         params: params,
         config: config)
       context[:gem_name] = 'aws-sdk-quicksight'
-      context[:gem_version] = '1.72.0'
+      context[:gem_version] = '1.72.1'
       Seahorse::Client::Request.new(handlers, context)
     end