+++ gems/aws-sdk-transcribeservice/lib/aws-sdk-transcribeservice.rb
@@ -52,6 +52,6 @@ require_relative 'aws-sdk-transcribeservice/customizations'
 # @!group service
 module Aws::TranscribeService
 
-  GEM_VERSION = '1.78.0'
+  GEM_VERSION = '1.79.0'
 
 end